refactor: extract circom dependencies into workspace extensions
- Move build-circuit and compiler binaries into separate workspace crates under extensions/
- Remove direct circom dependencies from main crate to reduce coupling
- Create workspace structure with extensions/build-circuit and extensions/compiler
- Move circom-specific type conversions (OperatorType, StatusInput) into build-circuit
- Relocate IODef and TemplateInstanceIOMap types to storage module
- Update build scripts to build entire workspace
- Add package metadata (authors, description, license) to main Cargo.toml
