+++ extensions/compiler/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "compiler"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+compiler = { git = "https://github.com/olomix/circom.git", branch = "master" }
+constraint_generation = { git = "https://github.com/olomix/circom.git", branch = "master" }
+program_structure = { git = "https://github.com/olomix/circom.git", branch = "master" }
+ruint = { version = "1.13.1", features = ["rand", "serde", "num-traits"] }
+type_analysis = { git = "https://github.com/olomix/circom.git", branch = "master" }
+circom-witnesscalc = { path = "../.." }
+parser = { git = "https://github.com/olomix/circom.git", branch = "master" }
